requirements.txt
setuptools~=57.4.0
gitex~=0.1
Pillow~=8.3.1